+++ examples/cis2/balanceOf.ts
@@ -1,7 +1,7 @@
 import { createConcordiumClient, CIS2Contract } from '@concordium/node-sdk';
 import { credentials } from '@grpc/grpc-js';
 import meow from 'meow';
-import { parseAddress } from '../shared/util';
+import { parseAddress, parseEndpoint } from '../shared/util';
 
 const cli = meow(
     `
@@ -49,16 +49,12 @@ const cli = meow(
     }
 );
 
-const [address, port] = cli.flags.endpoint.split(':');
+const [address, port] = parseEndpoint(cli.flags.endpoint);
 const client = createConcordiumClient(
     address,
     Number(port),
     credentials.createInsecure()
 );
-
-if (cli.flags.h) {
-    cli.showHelp();
-}
 
 (async () => {
     const contract = await CIS2Contract.create(client, {