+++ examples/client/getBlockSpecialEvents.ts
@@ -1,3 +1,4 @@
+import { parseEndpoint } from '../shared/util';
 import {
     BlockSpecialEvent,
     createConcordiumClient,
@@ -22,7 +23,6 @@ const cli = meow(
             block: {
                 type: 'string',
                 alias: 'b',
-                default: '', // This defaults to LastFinal
             },
             endpoint: {
                 type: 'string',
@@ -33,7 +33,8 @@ const cli = meow(
     }
 );
 
-const [address, port] = cli.flags.endpoint.split(':');
+const [address, port] = parseEndpoint(cli.flags.endpoint);
+
 const client = createConcordiumClient(
     address,
     Number(port),