+++ examples/common/cis0Supports.ts
@@ -5,6 +5,7 @@ import {
 } from '@concordium/node-sdk';
 import { credentials } from '@grpc/grpc-js';
 import meow from 'meow';
+import { parseEndpoint } from '../shared/util';
 
 const cli = meow(
     `
@@ -18,7 +19,7 @@ const cli = meow(
   Options
     --help,                 -h  Displays this message
     --endpoint,             -e  Specify endpoint of a grpc2 interface of a Concordium node in the format "address:port". Defaults to 'localhost:20000'
-    --subindex,                 The subindex of the smart contract. Defaults to 0
+    --subindex,             -s  The subindex of the smart contract. Defaults to 0
 `,
     {
         importMeta: import.meta,
@@ -35,6 +36,7 @@ const cli = meow(
             },
             subindex: {
                 type: 'number',
+                alias: 's',
                 default: 0,
             },
             supportsQuery: {
@@ -47,16 +49,12 @@ const cli = meow(
     }
 );
 
-const [address, port] = cli.flags.endpoint.split(':');
+const [address, port] = parseEndpoint(cli.flags.endpoint);
 const client = createConcordiumClient(
     address,
     Number(port),
     credentials.createInsecure()
 );
-
-if (cli.flags.h) {
-    cli.showHelp();
-}
 
 (async () => {
     const contractAddress: ContractAddress = {