+++ examples/common/delegationRemove.ts
@@ -1,0 +1,99 @@
+import {
+    AccountAddress,
+    AccountTransaction,
+    AccountTransactionHeader,
+    AccountTransactionType,
+    signTransaction,
+    TransactionExpiry,
+    createConcordiumClient,
+    ConfigureDelegationPayload,
+    CcdAmount,
+    parseWallet,
+    buildAccountSigner,
+} from '@concordium/node-sdk';
+import { credentials } from '@grpc/grpc-js';
+import { readFileSync } from 'node:fs';
+import { parseEndpoint } from '../shared/util';
+
+import meow from 'meow';
+
+const cli = meow(
+    `
+  Usage
+    $ yarn run-example <path-to-this-file> [options]
+
+  Required
+    --wallet-file, -w  A path to a wallet export file from a Concordium wallet
+
+  Options
+    --help,     -h  Displays this message
+    --endpoint, -e  Specify endpoint of the form "address:port", defaults to localhost:20000
+`,
+    {
+        importMeta: import.meta,
+        flags: {
+            walletFile: {
+                type: 'string',
+                alias: 'w',
+                isRequired: true,
+            },
+            endpoint: {
+                type: 'string',
+                alias: 'e',
+                default: 'localhost:20000',
+            },
+        },
+    }
+);
+
+const [address, port] = parseEndpoint(cli.flags.endpoint);
+const client = createConcordiumClient(
+    address,
+    Number(port),
+    credentials.createInsecure()
+);
+
+/**
+ * The following example demonstrates how an account can be configured to
+ * not be a delegator anymore.
+ */
+
+(async () => {
+    // Read wallet-file
+    const walletFile = readFileSync(cli.flags.walletFile, 'utf8');
+    const wallet = parseWallet(walletFile);
+    const sender = new AccountAddress(wallet.value.address);
+
+    const header: AccountTransactionHeader = {
+        expiry: new TransactionExpiry(new Date(Date.now() + 3600000)),
+        nonce: (await client.getNextAccountNonce(sender)).nonce,
+        sender: sender,
+    };
+
+    const configureDelegationPayload: ConfigureDelegationPayload = {
+        stake: new CcdAmount(0n),
+    };
+
+    const configureDelegationTransaction: AccountTransaction = {
+        header: header,
+        payload: configureDelegationPayload,
+        type: AccountTransactionType.ConfigureDelegation,
+    };
+
+    // Sign transaction
+    const signer = buildAccountSigner(wallet);
+    const signature = await signTransaction(
+        configureDelegationTransaction,
+        signer
+    );
+
+    const transactionHash = await client.sendAccountTransaction(
+        configureDelegationTransaction,
+        signature
+    );
+
+    console.log('Transaction submitted, waiting for finalization...');
+
+    const status = await client.waitForTransactionFinalization(transactionHash);
+    console.dir(status, { depth: null, colors: true });
+})();