+++ examples/shared/util.ts
@@ -8,3 +8,12 @@ export const parseAddress = (input: string): Base58String | ContractAddress => {
     const [i, si] = input.split(',');
     return { index: BigInt(i), subindex: BigInt(si) };
 };
+
+export const parseEndpoint = (endpoint: string): [string, number] => {
+    // Split endpoint on last colon
+    const lastColonIndex = endpoint.lastIndexOf(':');
+    const address = endpoint.substring(0, lastColonIndex);
+    const port = Number(endpoint.substring(lastColonIndex + 1));
+
+    return [address, port];
+};